change: move modelcourse components
